+++ src/core/enums/AppFeatureStatus.ts
@@ -1,0 +1,10 @@
+export enum AppFeatureStatus {
+	// generally activated
+	ACTIVATED = 'activated',
+	// not available
+	DEACTIVATED = 'deactivated',
+	// not in production
+	TESTING = 'testing',
+	// only for admins
+	ADMINISTRATIVE = 'admistrative',
+}