+++ src/core/index.ts
@@ -2,4 +2,5 @@ export * from './router/auth.router';
 export * from './guards/auth.guard';
 export * from './controllers/app-websocket.controller';
 export * from './controllers/app-api.controller';
+export * from './services/app-feature.service';
 export * from './errors';